core/translate: move INSERT statement binding into the bind module
INSERT was the one DML with no statement-level bind entry point: target
validation, VALUES default resolution, single-row VALUES binding, and
UPSERT conflict-target resolution all lived inside translate_insert.
Move them into bind::bind_insert_stmt, which returns a BoundInsert
carrying the validated table and database id. Body binding stays
intentionally partial (single-row VALUES bind scope-less, UPSERT DO
UPDATE defers to emission for EXCLUDED, SELECT sources bind when the
source is planned, virtual tables consume the unbound body), but the
code now lives in one place.

Also bind recursive-CTE body-level LIMIT/OFFSET at bind time instead of
during recursive-CTE planning, which removes parse_limit and the last
bind_fixed_scope_expr call in the planner.
